+++ lib/logstash/plugin_mixins/aws_config/v2.rb
@@ -11,13 +11,15 @@ module LogStash::PluginMixins::AwsConfig::V2
   def aws_options_hash
     opts = {}
 
-    if @access_key_id.is_a?(NilClass) ^ @secret_access_key.is_a?(NilClass)
-      @logger.warn("Likely config error: Only one of access_key_id or secret_access_key was provided but not both.")
-    end
-
-    opts[:credentials] = credentials if credentials
-
     opts[:http_proxy] = @proxy_uri if @proxy_uri
+
+    if @role_arn
+      credentials = assume_role(opts)
+      opts = { :credentials => credentials }
+    else
+      credentials = aws_credentials
+      opts[:credentials] = credentials if credentials
+    end
 
     if self.respond_to?(:aws_service_endpoint)
       # used by CloudWatch to basically do the same as bellow (returns { region: region })
@@ -27,42 +29,45 @@ module LogStash::PluginMixins::AwsConfig::V2
       opts[:region] = @region
     end
 
-    if !@endpoint.is_a?(NilClass)
-      opts[:endpoint] = @endpoint
-    end
+    opts[:endpoint] = @endpoint unless @endpoint.nil?
 
     return opts
   end
 
   private
-  def credentials
-    @creds ||= begin
-                 if @access_key_id && @secret_access_key
-                   credentials_opts = {
-                     :access_key_id => @access_key_id,
-                     :secret_access_key => @secret_access_key.value
-                   }
 
-                   credentials_opts[:session_token] = @session_token.value if @session_token
-                   Aws::Credentials.new(credentials_opts[:access_key_id],
-                                        credentials_opts[:secret_access_key],
-                                        credentials_opts[:session_token])
-                 elsif @aws_credentials_file
-                   credentials_opts = YAML.load_file(@aws_credentials_file)
-                   Aws::Credentials.new(credentials_opts[:access_key_id],
-                                        credentials_opts[:secret_access_key],
-                                        credentials_opts[:session_token])
-                 elsif @role_arn
-                   assume_role
-                 end
-               end
+  def aws_credentials
+    if @access_key_id && @secret_access_key
+      Aws::Credentials.new(@access_key_id, @secret_access_key.value, @session_token ? @session_token.value : nil)
+    elsif @access_key_id.nil? ^ @secret_access_key.nil?
+      @logger.warn("Likely config error: Only one of access_key_id or secret_access_key was provided but not both.")
+      secret_access_key = @secret_access_key ? @secret_access_key.value : nil
+      Aws::Credentials.new(@access_key_id, secret_access_key, @session_token ? @session_token.value : nil)
+    elsif @aws_credentials_file
+      credentials_opts = YAML.load_file(@aws_credentials_file)
+      credentials_opts.default_proc = lambda { |hash, key| hash.fetch(key.to_s, nil) }
+      Aws::Credentials.new(credentials_opts[:access_key_id],
+                           credentials_opts[:secret_access_key],
+                           credentials_opts[:session_token])
+    else
+      nil # AWS client will read ENV or ~/.aws/credentials
+    end
   end
+  alias credentials aws_credentials
 
-  def assume_role
+  def assume_role(opts = {})
+    unless opts.key?(:credentials)
+      credentials = aws_credentials
+      opts[:credentials] = credentials if credentials
+    end
+
+    # for a regional endpoint :region is always required by AWS
+    opts[:region] = @region
+
     Aws::AssumeRoleCredentials.new(
-      :client => Aws::STS::Client.new(:region => @region),
-      :role_arn => @role_arn,
-      :role_session_name => @role_session_name
+        :client => Aws::STS::Client.new(opts),
+        :role_arn => @role_arn,
+        :role_session_name => @role_session_name
     )
   end
 end